Adds part editing Fixes a backend bug in updating parts Refactors a bunch of things including loader behavior, js source locations

+++ Binner/Binner.Common/Models/Requests/UpdatePartRequest.cs
@@ -7,6 +7,6 @@ namespace Binner.Common.Models
         /// <summary>
         /// The part id
         /// </summary>
-        public int PartId { get; set; }
+        public long PartId { get; set; }
     }
 }
